+++ lib/issue-rename-handler.js
