+++ lib/templates/comment.js
