+++ lib/templates/index.js
