+++ lib/templates/issue.js
