+++ lib/templates/issueFromMerge.js
