+++ lib/templates/reopenClosed.js
