+++ lib/utils/check-for-body.js
