+++ lib/utils/check-for-duplicate-issue.js
