+++ lib/utils/chunk-diff.js
