+++ lib/utils/generate-assigned-to.js
