+++ lib/utils/generate-label.js
