+++ lib/utils/get-deets.js
