+++ lib/utils/parse-chunk.js
