+++ lib/utils/should-exclude-file.js
